+++ source/state_representation/include/state_representation/space/cartesian/CartesianPose.hpp
@@ -2,10 +2,12 @@
 
 #include "state_representation/space/cartesian/CartesianState.hpp"
 #include "state_representation/space/cartesian/CartesianTwist.hpp"
+#include "state_representation/space/cartesian/CartesianAcceleration.hpp"
 #include "state_representation/space/cartesian/CartesianWrench.hpp"
 
 namespace state_representation {
 class CartesianTwist;
+class CartesianAcceleration;
 class CartesianWrench;
 
 /**
@@ -23,7 +25,7 @@ public:
   Eigen::Matrix<double, 6, 1> get_twist() const = delete;
   const Eigen::Vector3d& get_linear_acceleration() const = delete;
   const Eigen::Vector3d& get_angular_acceleration() const = delete;
-  Eigen::Matrix<double, 6, 1> get_accelerations() const = delete;
+  Eigen::Matrix<double, 6, 1> get_acceleration() const = delete;
   const Eigen::Vector3d& get_force() const = delete;
   const Eigen::Vector3d& get_torque() const = delete;
   Eigen::Matrix<double, 6, 1> get_wrench() const = delete;
@@ -32,7 +34,7 @@ public:
   void set_twist(const Eigen::Matrix<double, 6, 1>& twist) = delete;
   void set_linear_acceleration(const Eigen::Vector3d& linear_acceleration) = delete;
   void set_angular_acceleration(const Eigen::Vector3d& angular_acceleration) = delete;
-  void set_accelerations(const Eigen::Matrix<double, 6, 1>& accelerations) = delete;
+  void set_acceleration(const Eigen::Matrix<double, 6, 1>& acceleration) = delete;
   void set_force(const Eigen::Vector3d& force) = delete;
   void set_torque(const Eigen::Vector3d& torque) = delete;
   void set_wrench(const Eigen::Matrix<double, 6, 1>& wrench) = delete;
@@ -167,6 +169,13 @@ public:
    * @return the current CartesianPose multiplied by the CartesianTwist given in argument
    */
   CartesianTwist operator*(const CartesianTwist& twist) const;
+
+  /**
+   * @brief Overload the * operator
+   * @param acceleration CartesianAcceleration to multiply with
+   * @return the current CartesianPose multiplied by the CartesianAcceleration given in argument
+   */
+  CartesianAcceleration operator*(const CartesianAcceleration& acceleration) const;
 
   /**
    * @brief Overload the * operator